% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigreadr-package.R
\docType{package}
\name{bigreadr-package}
\alias{bigreadr}
\alias{bigreadr-package}
\title{bigreadr: Read Large Text Files}
\description{
Read large text files by splitting them in smaller files.
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
