% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cbind_df}
\alias{cbind_df}
\title{Merge data frames}
\usage{
cbind_df(...)
}
\arguments{
\item{...}{Multiple data frames with the same observations in the same order.}
}
\value{
One merged data frame with the names of the first input data frame.
}
\description{
Merge data frames
}
\examples{
cbind_df(iris, iris)
}
