% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{my_rbind}
\alias{my_rbind}
\title{Merge data frames}
\usage{
my_rbind(...)
}
\arguments{
\item{...}{Multiple data frames with the same variables in the same order.}
}
\value{
One merged data frame with the names of the first input data frame.
}
\description{
Merge data frames
}
\examples{
my_rbind(iris, iris)
}
