% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlines-split.R
\name{nlines}
\alias{nlines}
\title{Number of lines}
\usage{
nlines(file)
}
\arguments{
\item{file}{Path of the file.}
}
\value{
The number of lines as one integer.
}
\description{
Get the number of lines of a file.
}
\examples{
tmp <- fwrite2(iris)
nlines(tmp)

}
