% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{rbind_df}
\alias{rbind_df}
\title{Merge data frames}
\usage{
rbind_df(...)
}
\arguments{
\item{...}{Multiple data frames with the same variables in the same order.}
}
\value{
One merged data frame with the names of the first input data frame.
}
\description{
Merge data frames
}
\examples{
rbind_df(iris, iris)
}
