% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_file}
\alias{split_file}
\title{Split file}
\usage{
split_file(file, every_x_mb = 100, split = getOption("bigreadr.split"))
}
\arguments{
\item{file}{Path to file.}

\item{every_x_mb}{Maximum size of new files (default is \code{100} Mo).}

\item{split}{Path to system command 'split'. See README.}
}
\value{
A vector of paths to the new files.
}
\description{
Split file
}
