% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{big_fread}
\alias{big_fread}
\title{Read large text file}
\usage{
big_fread(file, every_nlines, .transform = identity,
  .combine = my_rbind, skip = 0, ..., print_timings = TRUE)
}
\arguments{
\item{file}{Path to file that you want to split.}

\item{every_nlines}{Maximum number of lines in new file parts.}

\item{.transform}{Function to transform each data frame corresponding to each
part of the \code{file}. Default doesn't change anything.}

\item{.combine}{Function to combine results. Should accept multiple arguments
(\code{...}) such as \code{rbind} (the default).}

\item{skip}{Number of lines to skip at the beginning of \code{file}.}

\item{...}{Other arguments to be passed to \link[data.table:fread]{data.table::fread},
excepted \code{input}, \code{file}, \code{skip} and \code{col.names}.}

\item{print_timings}{Whether to print timings? Default is \code{TRUE}.}
}
\value{
A vector of paths to the new files.
}
\description{
Read large text file by flitting it before.
}
