% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{big_fread}
\alias{big_fread}
\title{Read large text file}
\usage{
big_fread(file, .transform, .combine, skip = 0, ..., every_nlines = NULL,
  nlines = NULL, split = "split")
}
\arguments{
\item{file}{Path to file.}

\item{.transform}{Function to transform each part. See examples.}

\item{.combine}{Function to combine results.
Should accept multiple arguments (\code{...}) such as \code{rbind}.}

\item{skip}{Number of lines to skip at the beginning of \code{file}.}

\item{...}{Other arguments to be passed to \link[data.table:fread]{data.table::fread},
expected \code{input}, \code{file}, \code{skip} and \code{col.names}.}

\item{every_nlines}{Number of lines per new file.
Default splits the file in files of 500 MB.}

\item{nlines}{Number of lines of \code{file} (including possible header).
Default uses \code{fpeek::peek_count_lines(file)} to count the number of lines.
This argument is only used to guess \code{every_nlines}.}

\item{split}{Path to system command 'split'. See README.}
}
\value{
A vector of paths to the new files.
}
\description{
Read large text file by flitting it before.
}
