% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigreadr-package.R
\docType{package}
\name{bigreadr-package}
\alias{bigreadr}
\alias{bigreadr-package}
\title{bigreadr: Read Large Text Files}
\description{
Read large text files by splitting them in smaller files.
    Package 'bigreadr' also provides some convenient wrappers around fread()
    and fwrite() from package 'data.table'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/privefl/bigreadr}
  \item Report bugs at \url{https://github.com/privefl/bigreadr/issues}
}

}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
