% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{cbind_df}
\alias{cbind_df}
\title{Merge data frames}
\usage{
cbind_df(list_df)
}
\arguments{
\item{list_df}{A list of multiple data frames with the same observations in
the same order.}
}
\value{
One merged data frame.
}
\description{
Merge data frames
}
\examples{
cbind_df(list(iris, iris))
}
