% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{rbind_df}
\alias{rbind_df}
\title{Merge data frames}
\usage{
rbind_df(list_df)
}
\arguments{
\item{list_df}{A list of multiple data frames with the same variables in the
same order.}
}
\value{
One merged data frame with the names of the first input data frame.
}
\description{
Merge data frames
}
\examples{
str(iris)
str(rbind_df(list(iris, iris)))

}
