% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_file}
\alias{split_file}
\title{Split file}
\usage{
split_file(file, every_nlines = NULL, nlines = NULL, split = "split")
}
\arguments{
\item{file}{Path to file.}

\item{every_nlines}{Number of lines per new file.
Default splits the file in files of 500 MB.}

\item{nlines}{Number of lines of \code{file} (including possible header).
Default uses \code{fpeek::peek_count_lines(file)} to count the number of lines.
This argument is only used to guess \code{every_nlines}.}

\item{split}{Path to system command 'split'. See README.}
}
\value{
A vector of paths to the new files.
}
\description{
Split file
}
